nginx: terminate HTTP/3 connection on poll errors
This includes for example failures to parse HTTP/3 frames, which should
cause the connection to be shut down with an error.

In practice quiche should already be initiating connection close, but
due to bugs or other reasons this might not happen, so this change
provides an additional layer of robustness.

Signed-off-by: Diab Neiroukh <lazerl0rd@thezest.dev>